requirements.txt
python-telegram-bot==20.*
opencv-python-headless
python-dotenv
psycopg2-binary
